manage tickets

+++ app/Http/Controllers/TicketController.php
@@ -3,16 +3,19 @@
 namespace App\Http\Controllers;
 
 use App\Http\Consts\Ticketconsts;
+use App\Http\Consts\Userconsts;
+use App\Jobs\ticketMailjob;
+use App\Mail\Ticketmail;
 use App\Models\Response;
+use App\Models\Subject;
 use App\Models\Ticket;
 use App\Models\User;
 use Carbon\Carbon;
-use Database\Factories\UserFactory;
 use Exception;
 use Illuminate\Http\RedirectResponse as RedirectResponseAlias;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Mail;
 use UxWeb\SweetAlert\SweetAlert;
 
 class TicketController extends Controller
@@ -25,18 +28,18 @@ class TicketController extends Controller
 
     public function __construct()
     {
-        if (!auth()->user()) {
-            $this->middleware('auth');
-        }
+
+
     }
 
 
     public function index()
     {
         $user = auth()->user();
-        $tickets = Ticket::withTrashed()-> where('parent_id', '!=', 0)->where('user_id', $user->id)->orderby('subject')->latest()->paginate(2);
+        $tickets = Ticket::withTrashed()->where('user_id', $user->id)->orderby('created_at', 'DESC')->latest()->paginate(2);
         return view('ticket.show', compact('tickets'));
     }
+
 
     /**
      * Show the form for creating a new resource.
@@ -45,9 +48,10 @@ class TicketController extends Controller
      */
     public function create()
     {
-        $subjects = Ticket::where('parent_id', 0)->get();
-        $user = auth()->user();
-        return view('ticket.create', compact('user', 'subjects'));
+
+        $this->authorize('create', Ticket::class);
+        $subjects = Subject::all();
+        return view('ticket.create', compact('subjects'));
     }
 
 
@@ -57,9 +61,10 @@ class TicketController extends Controller
      */
     public function store(Request $request): RedirectResponseAlias
     {
+
         $request->validate([
             'title' => 'required',
-            'subject_parent' => 'required',
+            'subject' => 'required',
             'description' => 'required',
             'attachment.*' => 'nullable|mimes:jpg,jpeg,png,svg,pdf,txt',
         ]);
@@ -67,15 +72,17 @@ class TicketController extends Controller
             $file_name = create_name($request->attachment->getclientoriginalname());
             $request->attachment->move(public_path(env('UPLOAD_FILE')), $file_name);
         }
-        $subject=Ticket::find($request->subject_parent)->first()->subject;
-        Ticket::create([
-            'user_id' => $request->query('user'),
-            'parent_id' => $request->subject_parent,
-            'subject' =>  $subject,
+        $ticket = Ticket::create([
+            'user_id' => auth()->user()->id,
+            'subject_id' => $request->subject,
             'title' => $request->title,
             'description' => $request->description,
             'attachment' => $file_name ?? null,
         ]);
+        $users = User::where('role', Userconsts::POSHTIBAN)->get();
+        foreach ($users as $user) {
+            TicketMailJob::dispatchSync($ticket, $user);
+        }
         alert()->success('تیکت با موفقیت ایجاد گردید');
         return redirect()->route('ticket.index');
     }
@@ -88,6 +95,7 @@ class TicketController extends Controller
      */
     public function show(Ticket $ticket)
     {
+        $this->authorize('view', $ticket);
         return view('admin.show_ticket', compact('ticket'));
     }
 
@@ -99,7 +107,8 @@ class TicketController extends Controller
      */
     public function edit(Ticket $ticket)
     {
-        return view('admin.edit_ticket', compact('ticket'));
+        $this->authorize('update', $ticket);
+        return view('admin.edit_ticket&response', compact('ticket'));
 
     }
 
@@ -113,12 +122,12 @@ class TicketController extends Controller
     public function update(Request $request, Ticket $ticket)
     {
         $request->validate([
-            'response'=>'required',
+            'response.*' => 'required',
         ]);
         Response::find(key($request->response))->update([
-            'description'=>implode($request->response),
-            'user_id'=>auth()->user()->id,
-            'updated_at'=>Carbon::now(),
+            'description' => implode($request->response),
+            'user_id' => auth()->user()->id,
+            'updated_at' => Carbon::now(),
         ]);
         alert()->success('پاسخ تیکت ویرایش گردید');
         return redirect()->route('admin.index');
@@ -133,13 +142,14 @@ class TicketController extends Controller
      */
     public function destroy(Ticket $ticket)
     {
+        $this->authorize('delete', $ticket);
         try {
             DB::beginTransaction();
             $ticket->delete();
             $ticket->update([
-                'status'=> Ticketconsts::OPEN,
+                'status' => Ticketconsts::OPEN,
             ]);
-            foreach ($ticket->responses_methode as $response) {
+            foreach ($ticket->responses as $response) {
                 $response->update([
                     'deleted_at' => Carbon::now()
                 ]);
